Avoid unnecessary casting

+++ src/main/java/ca/ciccc/Assignment1.java
@@ -24,7 +24,7 @@ public class Assignment1 {
      * @return "1000.0 inches is 25.4 meters"
      */
     public static String inchesToMeters(int inches) {
-        return String.format("%d.0 inches is %.1f meters", inches, (float)(inches * 0.0254));
+        return String.format("%d.0 inches is %.1f meters", inches, inches * 0.0254);
     }
 
     /**
